third_party/blink/web_tests/external/wpt/WebCryptoAPI/import_export/Argon2_importKey.tentative.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2i}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2d}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 128 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 2: 3, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 192 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 3: 4, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey, deriveBits])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Good parameters: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [deriveBits, deriveKey, deriveBits, deriveKey])
  assert_unreached: Threw an unexpected error: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument Reached unreachable code
[FAIL] Empty Usages: 256 bits (raw-secret, {0: 1, 1: 2, 10: 11, 11: 12, 12: 13, 13: 14, 14: 15, 15: 16, 16: 17, 17: 18, 18: 19, 19: 20, 2: 3, 20: 21, 21: 22, 22: 23, 23: 24, 24: 25, 25: 26, 26: 27, 27: 28, 28: 29, 29: 30, 3: 4, 30: 31, 31: 32, 4: 5, 5: 6, 6: 7, 7: 8, 8: 9, 9: 10}, {name: Argon2id}, false, [])
  assert_equals: Should throw correct error, not TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Invalid keyFormat argument expected "SyntaxError" but got "TypeError"
Harness: the test ran to completion.
